Moved Landing page to presentational
